+++ nano/core_test/socket.cpp
@@ -1,5 +1,6 @@
 #include <nano/boost/asio/ip/address_v6.hpp>
 #include <nano/boost/asio/ip/network_v6.hpp>
+#include <nano/lib/thread_runner.hpp>
 #include <nano/lib/threading.hpp>
 #include <nano/node/transport/socket.hpp>
 #include <nano/test_common/system.hpp>
