+++ nano/nano_node/daemon.cpp
@@ -1,6 +1,7 @@
 #include <nano/boost/process/child.hpp>
 #include <nano/lib/signal_manager.hpp>
 #include <nano/lib/stacktrace.hpp>
+#include <nano/lib/thread_runner.hpp>
 #include <nano/lib/threading.hpp>
 #include <nano/lib/tlsconfig.hpp>
 #include <nano/lib/utility.hpp>
