+++ nano/nano_wallet/entry.cpp
@@ -3,7 +3,7 @@
 #include <nano/lib/cli.hpp>
 #include <nano/lib/errors.hpp>
 #include <nano/lib/rpcconfig.hpp>
-#include <nano/lib/threading.hpp>
+#include <nano/lib/thread_runner.hpp>
 #include <nano/lib/tlsconfig.hpp>
 #include <nano/lib/tomlconfig.hpp>
 #include <nano/lib/utility.hpp>
