+++ nano/node/ipc/ipc_server.cpp
@@ -5,6 +5,7 @@
 #include <nano/lib/config.hpp>
 #include <nano/lib/ipc.hpp>
 #include <nano/lib/locks.hpp>
+#include <nano/lib/thread_runner.hpp>
 #include <nano/lib/threading.hpp>
 #include <nano/lib/timer.hpp>
 #include <nano/node/ipc/action_handler.hpp>
